Cast GetGameTime to uint32

+++ CMangos/GlobalMethods.h
@@ -170,7 +170,7 @@ namespace LuaGlobalFunctions
      */
     int GetGameTime(Eluna* E)
     {
-        E->Push(eWorld->GetGameTime());
+        E->Push(uint32(eWorld->GetGameTime()));
         return 1;
     }
